% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.get_alnstats_for_bam}
\alias{.get_alnstats_for_bam}
\title{Internal function to get seqlengths, mapped and unmapped counts for a bam file}
\usage{
.get_alnstats_for_bam(bf)
}
\arguments{
\item{bf}{Character scalar with the path and filename to a bam file (index
is assumed to be already existing)}
}
\value{
a named numeric vector with elments 'seqlength', 'mapped' and 'unmapped'
}
\description{
Internal function to get seqlengths, mapped and unmapped counts for a bam file
}
\keyword{internal}
