% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qProject-class.R
\name{qProjectUpdate}
\alias{qProjectUpdate}
\title{Update qProject}
\usage{
qProjectUpdate(proj, quiet = TRUE)
}
\arguments{
\item{proj}{a \code{\link[=qProject-class]{qProject}} object.}

\item{quiet}{logical, whether to suppress update messages.}
}
\value{
A \code{\link[=qProject-class]{qProject}} object.
}
\description{
Check if a qProject is compatible with the current class definition, and 
update it if necessary.
}
\seealso{
\code{\link[=qProject-class]{qProject}}
}
\author{
Charlotte Soneson
}
\keyword{misc}
\keyword{utilities}
