% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuasR-package.R
\docType{package}
\name{QuasR-package}
\alias{QuasR-package}
\alias{QuasR}
\title{Quantify and Annotate Short Reads in R}
\description{
This package provides a pipeline for the quantification and analysis of 
short reads.
}
\details{
See packageDescription('QuasR') for package details.
}
\examples{
\dontrun{
# see qCount, qMeth and qProfile manual pages for examples
example(qCount)
example(qMeth)
example(qProfile)
}
}
\seealso{
\code{\link{qAlign}}, \code{\link{qCount}}, \code{\link{qProfile}}, 
  \code{\link{qMeth}}, \code{\link{qQCReport}}
}
\author{
Anita Lerch, Charlotte Soneson, Dimos Gaidatzis and Michael Stadler
}
\keyword{package}
