% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{worker_message}
\alias{worker_message}
\title{Build a formatted message for outputting to log files}
\usage{
worker_message(..., sep = "", appendLF = TRUE)
}
\description{
Build a formatted message for outputting to log files
}
\keyword{internal}
