\name{qProjectUpdate}
\alias{qProjectUpdate}
\title{Update qProject}
\description{
  Check if a qProject is compatible with the current class definition, and 
  update it if necessary. }
\usage{
  qProjectUpdate(proj, quiet=TRUE)
}
\arguments{
  \item{proj}{a \code{\link[=qProject-class]{qProject}} object.}
  \item{quiet}{logical, whether to suppress update messages.}
}
\details{}
\value{
  A \code{\link[=qProject-class]{qProject}} object.
}
%\references{
  %% ~put references to the literature/web site here ~
    %}
\author{
  Charlotte Soneson
}
%\note{
  %%  ~~further notes~~
    %}

%% ~Make other sections like Warning with \section{Warning }{....} ~
  
  \seealso{
    \code{\link[=qProject-class]{qProject}}
  }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
\keyword{misc}% __ONLY ONE__ keyword per line
